Rename CurrencyRateHNBProvider to ExchangeRateClient
Propagate the change throughout the code.
